# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HealthAI Coach API",
        "description": "Documentation de l'API HealthAI Coach, gérée avec Lomkit REST API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/users/search": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Rechercher et filtrer des utilisateurs",
                "operationId": "8150f61b3494506d102f807dbe12033a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/mutate": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer ou mettre à jour un/des utilisateurs",
                "operationId": "4028bf38fe6e02bf314c10451faf77d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mutate": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/destroy": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Supprimer un/des utilisateurs",
                "operationId": "8b20bb27e315355d4d06e001bfd552d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "resources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/restore": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Restaurer un/des utilisateurs supprimés",
                "operationId": "4d581fcc5ffdbf9082740024c4844f9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "resources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token Sanctum ici",
                "scheme": "bearer"
            }
        }
    },
...
}